+++ recalc-vscode/src/extension/extension.ts
@@ -6,6 +6,7 @@ import { ChildProcessWithoutNullStreams, execSync, spawn } from 'child_process';
 import { Client, MessageTransports } from '../rpc/client';
 import { SpreadsheetEditorProvider } from './spreadsheet-editor';
 import { existsSync } from 'fs';
+import { ExtensionLogger } from './logging';
 
 export function activate(context: vscode.ExtensionContext) {
 	// read vscode app-specific settings (as specified in package.json)
@@ -13,23 +14,28 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// json-rpc client implementation
 	const client = new class extends Client<SpreadsheetProtocol> {
-		public name = "recalc";
+		public readonly name = "recalc";
 
 		private process?: ChildProcessWithoutNullStreams;
+
+		constructor() {
+			super(config.serverMaxRestartCount, new ExtensionLogger("recalc"));
+		}
 
 		protected async createMessageTransports(encoding: "utf-8" | "ascii"): Promise<MessageTransports> {
 
 			// vsix is bundled with the -server-exe here:
-			const defaultBinPath = `${context.extensionPath}/bin/${this.name}-server-exe`;
+			const _binName = `${this.name}-server-exe`;
+			const defaultBinPath = `${context.extensionPath}/bin/${_binName}`;
 
 			// guess the location for the -server-exe
 			const binPath = config.serverPath
 				|| existsSync(defaultBinPath)
 						? defaultBinPath
-						: execSync("cabal list-bin recalc-server-exe").toString('utf-8').trim();
+						: execSync(`cabal list-bin ${_binName}`).toString('utf-8').trim();
 
 			// spawn and hook up message transports
-			this.log(`starting: ${binPath}`);
+			this.logger.log(`starting: ${binPath}`);
 			this.process = spawn("sh", ["-c", binPath]);
 
 			if (!this.process || !this.process.stdout || !this.process.stderr) {
@@ -39,18 +45,20 @@ export function activate(context: vscode.ExtensionContext) {
 			const reader = new rpc.StreamMessageReader(this.process.stdout, encoding);
 			const writer = new rpc.StreamMessageWriter(this.process.stdin, encoding);
 
-			this.process.stderr.on('data', data => this.info(`server-stderr: ${data.toString().trim()}`));
+			const binName = binPath.split("/").at(-1);
+			this.process.stderr.on('data', data => this.logger.info(`(${binName}) ${data.toString().trim()}`));
 
 			return {reader: reader, writer: writer}
 		}
 
 		override dispose(): void {
 			super.dispose();
+			this.logger.dispose?.();
 			if (this.process?.pid !== undefined) {
 				this.process.kill()
 			}
 		}
-	}(config.serverMaxRestartCount);
+	};
 
 	// register the custom editor and pass it the client instance
 	context.subscriptions.push(SpreadsheetEditorProvider.register(context, config, client))