+++ recalc-vscode/src/rpc/logging.ts
@@ -1,27 +1,37 @@
-/* eslint @typescript-eslint/no-namespace: 0 */
+/* eslint-disable  @typescript-eslint/no-explicit-any */
 
-export namespace Loglevel {
-	/**
-	 * An error message.
-	 */
-	export const Error = 1;
-	/**
-	 * A warning message.
-	 */
-	export const Warning = 2;
-	/**
-	 * An information message.
-	 */
-	export const Info = 3;
-	/**
-	 * A log message.
-	 */
-	export const Log = 4;
-	/**
-	 * A debug message.
-	 *
-	 */
-	export const Debug = 5;
+export enum Loglevel {
+  Error = 1,
+  Warning = 2,
+  Info = 3,
+  Debug = 4,
 }
 
-export type Loglevel = 1 | 2 | 3 | 4 | 5;
+/** extends rpc.Logger (compatible) with data and flag (when set vscode will open a notification) */
+export type IExtensionLogger = {
+	error: (message: string, error?: Error, data?: any, showNotification?: boolean) => void;
+	warn: (message: string, data?: any, showNotification?: boolean) => void;
+	info: (message: string, data?: any, showNotification?: boolean) => void;
+	log: (message: string, data?: any, showNotification?: boolean) => void;
+	dispose?: () => void;
+}
+
+/** use this Logger for testing */
+export const simpleLogger: IExtensionLogger = {
+	error: (message: string, error?: Error, data?: any, _showNotification?: boolean) =>
+		simpleOutput(console.error, Loglevel.Error, `${message}${error ? `(${error})` : ''}`, data),
+	warn: (message: string, data?: any, _showNotification?: boolean) =>
+		simpleOutput(console.warn, Loglevel.Warning, message, data),
+	info: (message: string, data?: any, _showNotification?: boolean) =>
+		simpleOutput(console.info, Loglevel.Info, message, data),
+	log: (message: string, data?: any, _showNotification?: boolean) =>
+		simpleOutput(console.log, Loglevel.Debug, message, data),
+};
+
+function simpleOutput(printer: (...data: any) => void, type: Loglevel, message: string, data?: any) {
+	const level = Loglevel[type];
+	printer(level, message)
+	if (data !== null && data !== undefined) {
+		printer(level, JSON.stringify(data, null, 2));
+	}
+}