+++ recalc-vscode/CHANGELOG.md
@@ -1,12 +1,14 @@
 # Revision history for recalc-vscode
 
-## 0.0.5 -- 2025-02-08
+## 0.0.7 -- 2025-02-09
 
 * Initial version: create project structure for a custom editor.
 * Add generic rpc implementation, client implementation and logging facilities.
 * Integrate type-safe json-rpc communication between client and server.
 * Add simple UI code using [Univer core/sheet-api][univer-sheet-api].
 * Add test suite runner, package the extension with Nix.
+* Implement json-rpc tests.
+* Add client test: connect to server, send "open" message and check response.
 
 <!-- References -->